Create A CURD Operation Logic For Controller

+++ controller/courseController.js
@@ -4,12 +4,18 @@ export const addCourse = async (req, res) => {
     try {
         let { courseName } = req.body
 
-        const findCourse = await Course.find({ courseName })
+        const findCourse = await Course.findOne({ courseName })
         if (findCourse) {
             return res.status(200).json({ success: true, message: "alrday Extis Course" })
         }
-        const createCourse = await Course.create({ courseName })
-        res.status(201).json({ success: true, message: "Course Add SuccessFully..." })
+
+        if (req.user && req.user.role == "Admin") {
+            const createCourse = await Course.create({ courseName })
+            return res.status(201).json({ success: true, message: "Course Add SuccessFully..." })
+        } else {
+            return res.status(401).json({ success: false, message: "Unauthorized Page" });
+        }
+
     } catch (error) {
         console.log(`Error By Course Controller Js For addCourse`, error);
         res.status(500).json({ success: false, msg: 'Internal  Error.', Error: error })
@@ -30,12 +36,17 @@ export const updateCourse = async (req, res) => {
     try {
         let { id } = req.params
         let { courseName } = req.body
-        const findCourse = await Course.findById({ id })
+        const findCourse = await Course.findById(id)
         if (!findCourse) {
             return res.status(404).json({ success: false, message: "No Data Found.." })
         }
-        let Course = await Course.findByIdAndUpdate({ id, courseName })
-        return res.status(201).json({ success: true, message: "Update Course SuccessFully...", data: Course })
+        if (req.user && req.user.role == "Admin") {
+            let Course = await Course.findByIdAndUpdate({ id, courseName })
+            return res.status(201).json({ success: true, message: "Update Course SuccessFully...", data: Course })
+        } else {
+            return res.status(401).json({ success: false, message: "Unauthorized Page" });
+        }
+
 
     } catch (error) {
         console.log(`Error By Course Controller Js For updateCourse`, error);
@@ -54,10 +65,14 @@ export const deleteCourse = async (req, res) => {
         if (existingUsers.length === 0) {
             return res.status(404).json({ message: "No matching records found." });
         }
-
+        if(req.user && req.user.role == "Admin"){
+            const result = await Course.deleteMany({ _id: { $in: ids } });
+        return res.json({ message: "Deleted successfully", data: result });
+        }else{
+            return res.status(401).json({ success: false, message: "Unauthorized Page" });
+        }
         // Delete records
-        const result = await Course.deleteMany({ _id: { $in: ids } });
-        return res.json({ message: "Deleted successfully", data:result });
+        
     } catch (error) {
         console.log(`Error By Course Controller Js For deleteCourse`, error);
         res.status(500).json({ success: false, msg: 'Internal  Error.', Error: error })
